% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_model.R
\name{classifier}
\alias{classifier}
\title{K visualization}
\usage{
classifier(data, split = 5, gridFrom = 1, gridTo = 21)
}
\arguments{
\item{heart_training}{heart_training is a a dataset used to be plot accuracy versus K}

\item{split}{is the number of diagnosis}

\item{gridFrom}{is the gridvals value}

\item{gridTo}{is the gridvals value}
}
\value{
a line graph with the accuracy estimates versus the number of neighbors
}
\description{
K visualization
}
\examples{
classifier(heart_training, 5, 1, 21)
}
