% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Testing accuracy of model}
\usage{
confusion_matrix(data, x, y, hGraph = 7, wGraph = 7, textSize = 22)
}
\arguments{
\item{data}{is the predicted data of our classifier}

\item{x}{is diagnosis_f}

\item{y}{is .pred_Class}

\item{hGraph}{is the height of the Confusion matrix}

\item{wGraph}{is the width of the Confusion matrix}

\item{textSize}{is size of text}
}
\value{
a Confusion Matrix heat map with the accuracies based on variables of the model
}
\description{
Testing accuracy of model
}
\examples{
confusion_matrix(data_predict, 7, 7, 22)

}
