% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_forward_function.R
\name{forwardSelection}
\alias{forwardSelection}
\title{Forward selection step in model}
\usage{
forwardSelection(data, seedVar = 1, folds = 5)
}
\arguments{
\item{data}{a dataset that is intented to be turned into a bar graph}

\item{seedVar}{the number inputted into set.seed()}

\item{folds}{the number of folds for cross validation}
}
\value{
a table with the accuracies based on variables added to the model
}
\description{
Forward selection step in model
}
\examples{
forwardSelection(diff_data_set0)
forwardSelection(diff_data_set1, seedVar = 2, folds = 10)
}
