% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart_data_accuracies.R
\docType{data}
\name{heart_data_accuracies}
\alias{heart_data_accuracies}
\title{heart_data_accuracies}
\format{
A data frame with 21 rows and 7 variables:
}
\usage{
heart_data_accuracies
}
\description{
A dataset containing information the heart
}
\keyword{heart_data_accuracies}
