% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_visualization.R
\name{knn_visualization}
\alias{knn_visualization}
\title{knn Visualization}
\usage{
knn_visualization(data)
}
\arguments{
\item{data}{Data frame used for visualization.}
}
\value{
Visualization for Neighbors vs Accuracy Estimate.
}
\description{
knn Visualization
}
\examples{
knn_visualization(heart_data_accuracies)
}
