% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_classifier.R
\docType{data}
\name{majority_classifier}
\alias{majority_classifier}
\title{Majority Classifier}
\format{
A data frame with 205 rows and 26 variables:
}
\usage{
majority_classifier
}
\description{
A dataset containing information for the majority classifier visualization
}
\keyword{majority_classifier}
