% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_classifier_visualization.R
\name{majority_classifier_vis_function}
\alias{majority_classifier_vis_function}
\title{majority_classifier_vis_function}
\usage{
majority_classifier_vis_function(
  data,
  hGraph = 5,
  wGraph = 7,
  textSize = 20,
  colWidth = 0.5,
  xLabel = "Heart disease degree of severity",
  yLabel = "Percent of outcomes \\n in training dataset",
  titleLabel = "Label proportions of classifier in dataset"
)
}
\arguments{
\item{data}{Data frame used for visualization.}

\item{hGraph}{Height of the graph.}

\item{wGraph}{Width of the graph.}

\item{textSize}{Size of the text.}

\item{colWidth}{The width of the Column.}

\item{xLabel}{Label for the x axis.}

\item{yLabel}{Label for the y axis.}

\item{titleLabel}{Label for the title.}
}
\value{
Visualization for Neighbors vs Accuracy Estimate.
}
\description{
majority_classifier_vis_function
}
\examples{
majority_classifier_vis_function(majority_classifier)
}
