% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_split_new.R
\docType{data}
\name{training_split_new}
\alias{training_split_new}
\title{training_split_new}
\format{
A data frame with 403 rows and 3 variables:
}
\usage{
training_split_new
}
\description{
A dataset containing information the heart
}
\keyword{training_split_new}
