% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_boxplot.R
\name{grid_boxplot}
\alias{grid_boxplot}
\title{Make a grid boxplot}
\usage{
grid_boxplot(data, yAxis, yLabel, titleLabel)
}
\arguments{
\item{data}{The dataset we're using.}

\item{yAxis}{The Y Axis parameter.}

\item{yLabel}{The label for the Y Axis.}

\item{titleLabel}{The label for the title.}
}
\value{
A visualization of a boxplot.
}
\description{
Make a grid boxplot
}
\examples{
grid_boxplot(mtcars, age, "Age (years)",
"A. Boxplot of degree of heart \n disease in relation to patient's \nage")
}
