% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_visualization.R
\name{knn_visualization}
\alias{knn_visualization}
\title{kNN visualization}
\usage{
knn_visualization(data, x, y, xlab, ylab, title)
}
\arguments{
\item{data}{accuracy data created by kNN model
in classification_model.R}

\item{x}{is the horizontal variable for knn model}

\item{y}{is the vertical variable for knn model}

\item{xlab}{is the name of the x-axis}

\item{ylab}{is the name of the y-axis}

\item{title}{is the name of the model}
}
\value{
Elbow plot for kNN models.
}
\description{
kNN visualization
}
