% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_visualization.R
\name{knn_visualization}
\alias{knn_visualization}
\title{kNN visualization}
\usage{
knn_visualization(data)
}
\arguments{
\item{data}{accuracy data created by kNN model
in classification_model.R}
}
\value{
Elbow plot for kNN models.
}
\description{
kNN visualization
}
