% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_classifier_function.R
\name{majority_classifier_vis_function}
\alias{majority_classifier_vis_function}
\title{Converts dataset to a bar graph visualization}
\usage{
majority_classifier_vis_function(
  data,
  hGraph = 5,
  wGraph = 7,
  textSize = 20,
  colWidth = 0.5,
  xLabel = "Heart disease degree of severity",
  yLabel = "Percent of outcomes \\n in training dataset",
  titleLabel = "Label proportions of classifier in dataset"
)
}
\arguments{
\item{data}{a dataset that is intented to be turned into a bar graph}

\item{hGraph}{the height of the bar graph}

\item{wGraph}{the width of the bar graph}

\item{textSize}{the text size of the labels and title in the graph}

\item{colWidth}{the column width of the bars in the graph}

\item{xLabel}{the label on the x-axis}

\item{yLabel}{the label on the y-axis}

\item{titleLabel}{the title of the graph}
}
\value{
a visualization of the majority classifier
}
\description{
Converts dataset to a bar graph visualization
}
\examples{
majority_classifier_vis_function(diff_data_set0)
majority_classifier_vis_function(diff_data_set1, hgraph = 4, wGraph = 10, colWidth = 1)
}
